Some final cleanup and todo notes
